refactor:  컴포넌트명 변경
- ...Group -> ...s로 변경
