+++ frontend/src/pages/PollProgressPage/components/PollProgressForm/PollProgressForm.tsx
@@ -7,7 +7,7 @@ import Divider from '../../../../components/Divider/Divider';
 import MarginContainer from '../../../../components/MarginContainer/MarginContainer';
 import PollProgressButtons from '../PollProgressButtons/PollProgressButtons';
 import { getPoll, progressPoll, getPollItems } from '../../../../api/poll';
-import PollProgressItemGroup from '../PollProgressItemGroup/PollProgressItemGroup';
+import PollProgressItems from '../PollProgressItems/PollProgressItems';
 import {
   Poll,
   SelectedPollItem,
@@ -168,7 +168,7 @@ function PollProgressForm() {
             />
           </MarginContainer>
           <MarginContainer margin="0 0 15.2rem 0">
-            <PollProgressItemGroup
+            <PollProgressItems
               pollItems={pollItems}
               selectedPollItems={selectedPollItems}
               allowedPollCount={poll.allowedPollCount}
