+++ frontend/src/pages/PollProgressPage/components/PollProgressItems/PollProgressItems.styles.tsx
