+++ frontend/src/pages/PollResultPage/components/PollResultItems/PollResultItems.styles.tsx
